requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.0.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cpu
torchvision==0.16.1+cpu